moving bundle caching tests that use the watcher threads to the integration test suite
